[FIX] account_edi_ubl_cii,account_peppol: R010/R020 checks only when peppol checked
Some people are using BIS3 files even for B2C. They give the xml files to the
accountant for him/her to import the file and generate the invoice in the
accounting software. In that case, since it's a B2C transaction, the EndpointId
might not be set but that's ok since it's only used to generate the invoice.

opw-5952109
