+++ addons/account_edi_ubl_cii/models/account_edi_xml_ubl_bis3.py
@@ -777,21 +777,6 @@ class AccountEdiXmlUBLBIS3(models.AbstractModel):
                     "The VAT number of the supplier does not seem to be valid. It should be of the form: NO179728982MVA."
                 ) if not mva.is_valid(vat) or len(vat) != 14 or vat[:2] != 'NO' or vat[-3:] != 'MVA' else "",
             })
-
-        # [PEPPOL-EN16931-R010]
-        if not vals['document_node']['cac:AccountingCustomerParty']['cac:Party']['cbc:EndpointID']['_text']:
-            constraints['ubl_peppol_en16931-r010'] = _(
-                "[PEPPOL-EN16931-R010] An electronic address (EAS) must be provided on the customer '%s'.",
-                vals['customer'].display_name,
-            )
-
-        # [PEPPOL-EN16931-R020]
-        if not vals['document_node']['cac:AccountingSupplierParty']['cac:Party']['cbc:EndpointID']['_text']:
-            constraints['ubl_peppol_en16931-r020'] = _(
-                "[PEPPOL-EN16931-R020] An electronic address (EAS) must be provided on the company '%s'.",
-                vals['supplier'].display_name,
-            )
-
         return constraints
 
     # -------------------------------------------------------------------------
