+++ addons/account_peppol/models/__init__.py
@@ -1,6 +1,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import account_edi_proxy_user
+from . import account_edi_xml_ubl_bis3
 from . import account_journal
 from . import account_move
 from . import account_move_send
